Backend design changes
These changes store reference to the patches instead of storing entire patch
contents. Some patch metadata is stored in the database - the diffstat and the
comments.

There have been changes to the API as well - Patch metadata is retrieved
individually when called by the user. This helps making the requests small and
page loads considerably faster